requirements.txt
backports.tarfile==1.2.0
django-passkeys==1.2.7
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
packaging==24.2
platformdirs==4.2.2
tomli==2.0.1
